Ajustando problema com valor nulo no retorno da chamada de cartoes

+++ mscartoes/mscartoes/src/main/java/com/starking/mscartoes/domain/DadosSolicitacaoEmissaoCartao.java
@@ -1,14 +1,10 @@
 package com.starking.mscartoes.domain;
-import java.math.BigDecimal;
-
 import lombok.Data;
 
 @Data
 public class DadosSolicitacaoEmissaoCartao {
-
 	private Long idCartao;
 	private String cpf;
 	private String endereco;
-	private BigDecimal limiteLiberado;
-
+	private Double limiteLiberado;
 }
